feito category, cadastro de produto funcionando certo, precisa add forma de salvaar imagem

+++ composeApp/src/commonMain/kotlin/dev/azeredo/domain/repository/ProductRepository.kt
@@ -4,6 +4,6 @@ interface ProductRepository {
     suspend fun addProduct(product: Product)
     suspend fun removeProduct(product: Product)
     suspend fun updateProduct(product: Product)
-    suspend fun getProductById(productId: Int): Product?
+    suspend fun getProductById(productId: Long): Product?
     suspend fun getAllProducts(): Flow<List<Product>>
 }
